rename 'work item' to 'service item'
