+++ app/Models/ServiceItem.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class WorkItems extends Model
+class ServiceItem extends Model
 {
     use HasFactory;
