+++ app/Repositories/ServiceItemRepository.php
@@ -1,0 +1,11 @@
+<?php
+ 
+namespace App\Repositories;
+ 
+use App\Models\ServiceItem;
+ 
+class ServiceItemRepository
+{
+
+}
+
