--- app/Repositories/WorkItemRepository.php
@@ -1,11 +1,0 @@
-<?php
- 
-namespace App\Repositories;
- 
-use App\Models\WorkItems;
- 
-class WorkItemRepository
-{
-
-}
-
