+++ database/seeders/WorkItemsSeeder.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
-class WorkItemsSeeder extends Seeder
+class ServiceItemsSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -17,7 +17,7 @@ class WorkItemsSeeder extends Seeder
     public function run()
     {
         for($i = 0; $i < 10; $i++) { 
-            DB::table('work_items')->insert([
+            DB::table('service_items')->insert([
                 'service_date' => now()->format('Y-m-d H:i:s'),
                 'vehicle_id' => 32,
                 'mileage' => rand(30000,100000),
