+++ resources/js/components/AddServiceItemModalComponent.vue
@@ -7,52 +7,36 @@
     export default {
         setup () {
             return { 
-                v$: useVuelidate(),
+                v$: useVuelidate()
             }
         },
         data () {
             return {
-                results: [],
+                vehicleid: "",
+                date: "",
+                mileage: "",
+                service_summary: "",
+                service_details: "",
+                technician: "",
+                cost: ""
             }
         },
+
         // vuelidate validation rules
         validations () {
             return {
-                results: {
-                    service_date: {
-                        required
-                    },
-                    mileage: { 
-                        required
-                    },
-                    technician: {
-                        required
-                    },
-                    cost: {
-                        required
-                    },
-                    service_summary: {
-                        required
-                    },
-                }
+                date: { required },
+                mileage: { required },
+                service_summary: { required },
+                technician: { required },
+                cost: { required }
             }
         },
         components: {
             'editor': Editor
         },
-        mounted () {
-            let workItemID = this.$attrs.data.workItemID;
-            axios.get('../workitems/' + workItemID)
-            .then(response => {
-                // console.log(response)
-                this.results = response.data
-            })
-            .catch(error => console.log(error));
-        },
         methods: {
             async submitForm () {
-                let workItemID = this.$attrs.data.workItemID;
-                let vehicleID = this.$attrs.data.vehicleid;
                 const isFormCorrect = await this.v$.$validate()
 
                 // if validation returns errors, do nothing
@@ -60,19 +44,18 @@
 
                 // otherwise, submit form
                 // gather data from form
-                const workItemProps = {
-                    workItemID: workItemID,
-                    date: this.results.service_date, 
-                    mileage: this.results.mileage,
-                    service_summary: this.results.service_summary,
-                    service_details: (this.results.service_details) ? this.results.service_details : "n/a",
-                    technician: this.results.technician, 
-                    cost: this.results.cost
+                const serviceItemProps = {
+                    vehicle_id: this.$attrs.data.vehicleid,
+                    date: this.$refs.date.value, 
+                    mileage: this.$refs.mileage.value, 
+                    service_summary: this.$refs.service_summary.value,
+                    service_details: (this.$data.service_details) ? this.$data.service_details : "n/a",
+                    technician: this.$refs.technician.value, 
+                    cost: this.$refs.cost.value
                 }
 
                 // json the data from form
-                const data = JSON.stringify({workItemProps})
-                // console.log(data);
+                const data = JSON.stringify({serviceItemProps})
                 const config = {
                     headers: {
                         'Content-Type': 'application/json',
@@ -82,11 +65,11 @@
                 }
 
                 // submit data to laravel
-                axios.post('../workitems/' + workItemID, data, config)
+                axios.post('../serviceitems', data, config)
                     // .then(response => console.log(response))
                     .then(response => {
                         // console.log(response)
-                        window.location.href = '/vehicles/' + vehicleID;
+                        window.location.href = '/vehicles/' + serviceItemProps.vehicle_id;
                     })
                     .catch(error => console.log(error));
             }
@@ -101,7 +84,7 @@
                 <div class="card">
                     <div class="card-header">
                         <slot name="header">
-                            Edit Service Log Item
+                            Add a New Service Log Item
                         </slot>
                     </div>
 
@@ -110,39 +93,39 @@
                             <form action="submitForm" method="POST">
                                 <div class="mb-3">
                                     <label for="date">Date</label>
-                                    <input type="date" name="date" id="date" class="form-control" v-model="results.service_date" ref="date">
-                                    <div class="text-danger" v-if="v$.results.service_date.$error">Date field is required.</div>                              
+                                    <input type="date" name="date" id="date" class="form-control" v-model="date" ref="date">
+                                    <div class="text-danger" v-if="v$.date.$error">Date field is required.</div>                              
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="mileage">Mileage</label>
-                                    <input type="number" name="mileage" id="mileage" class="form-control" v-model="results.mileage" ref="mileage">
-                                    <div class="text-danger" v-if="v$.results.mileage.$error">Mileage field is required.</div>
+                                    <input type="number" name="mileage" id="mileage" class="form-control" v-model="mileage" ref="mileage">
+                                    <div class="text-danger" v-if="v$.mileage.$error">Mileage field is required.</div>
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="technician">Performed by</label>
-                                    <input type="text" name="technician" id="technician" class="form-control" v-model="results.technician" ref="technician">
-                                    <div class="text-danger" v-if="v$.results.technician.$error">Performed by field is required.</div>    
+                                    <input type="text" name="technician" id="technician" class="form-control" v-model="technician" ref="technician">
+                                    <div class="text-danger" v-if="v$.technician.$error">Performed by field is required.</div>    
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="cost">Cost</label>
-                                    <input type="number" step=0.01 name="cost" id="cost" class="form-control" v-model="results.cost" ref="cost">
-                                    <div class="text-danger" v-if="v$.results.cost.$error">Cost field is required.</div>    
+                                    <input type="number" step=0.01 name="cost" id="cost" class="form-control" v-model="cost" ref="cost">
+                                    <div class="text-danger" v-if="v$.cost.$error">Cost field is required.</div>    
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="service_summary">Service summary</label>
-                                    <input type="text" name="service_summary" id="service_summary" class="form-control" v-model="results.service_summary" ref="service_summary">
-                                    <div class="text-danger" v-if="v$.results.service_summary.$error">Service summary field is required.</div>                          
+                                    <input type="text" name="service_summary" id="service_summary" class="form-control" v-model="service_summary" ref="service_summary">
+                                    <div class="text-danger" v-if="v$.service_summary.$error">Service summary field is required.</div>                          
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="service_details">Service details</label>
                                     <editor 
                                         api-key="no-api-key"
-                                        v-model="results.service_details" 
+                                        v-model="service_details" 
                                         ref="service_details"
                                         :init="{
                                             height: 150,
@@ -160,7 +143,7 @@
                                             Cancel
                                         </button>
 
-                                        <button class="btn btn-primary" @click.prevent="submitForm">Update</button>
+                                        <button class="btn btn-primary" @click.prevent="submitForm">Submit</button>
                                     </slot>
                                 </div>
                             </form>
