+++ resources/js/components/EditServiceItemModalComponent.vue
@@ -7,36 +7,52 @@
     export default {
         setup () {
             return { 
-                v$: useVuelidate()
+                v$: useVuelidate(),
             }
         },
         data () {
             return {
-                vehicleid: "",
-                date: "",
-                mileage: "",
-                service_summary: "",
-                service_details: "",
-                technician: "",
-                cost: ""
+                results: [],
             }
         },
-
         // vuelidate validation rules
         validations () {
             return {
-                date: { required },
-                mileage: { required },
-                service_summary: { required },
-                technician: { required },
-                cost: { required }
+                results: {
+                    service_date: {
+                        required
+                    },
+                    mileage: { 
+                        required
+                    },
+                    technician: {
+                        required
+                    },
+                    cost: {
+                        required
+                    },
+                    service_summary: {
+                        required
+                    },
+                }
             }
         },
         components: {
             'editor': Editor
         },
+        mounted () {
+            let serviceItemID = this.$attrs.data.serviceItemID;
+            axios.get('../serviceitems/' + serviceItemID)
+            .then(response => {
+                // console.log(response)
+                this.results = response.data
+            })
+            .catch(error => console.log(error));
+        },
         methods: {
             async submitForm () {
+                let serviceItemID = this.$attrs.data.serviceItemID;
+                let vehicleID = this.$attrs.data.vehicleid;
                 const isFormCorrect = await this.v$.$validate()
 
                 // if validation returns errors, do nothing
@@ -44,18 +60,19 @@
 
                 // otherwise, submit form
                 // gather data from form
-                const workItemProps = {
-                    vehicle_id: this.$attrs.data.vehicleid,
-                    date: this.$refs.date.value, 
-                    mileage: this.$refs.mileage.value, 
-                    service_summary: this.$refs.service_summary.value,
-                    service_details: (this.$data.service_details) ? this.$data.service_details : "n/a",
-                    technician: this.$refs.technician.value, 
-                    cost: this.$refs.cost.value
+                const serviceItemProps = {
+                    serviceItemID: serviceItemID,
+                    date: this.results.service_date, 
+                    mileage: this.results.mileage,
+                    service_summary: this.results.service_summary,
+                    service_details: (this.results.service_details) ? this.results.service_details : "n/a",
+                    technician: this.results.technician, 
+                    cost: this.results.cost
                 }
 
                 // json the data from form
-                const data = JSON.stringify({workItemProps})
+                const data = JSON.stringify({serviceItemProps})
+                // console.log(data);
                 const config = {
                     headers: {
                         'Content-Type': 'application/json',
@@ -65,11 +82,11 @@
                 }
 
                 // submit data to laravel
-                axios.post('../workitems', data, config)
+                axios.post('../serviceitems/' + serviceItemID, data, config)
                     // .then(response => console.log(response))
                     .then(response => {
                         // console.log(response)
-                        window.location.href = '/vehicles/' + workItemProps.vehicle_id;
+                        window.location.href = '/vehicles/' + vehicleID;
                     })
                     .catch(error => console.log(error));
             }
@@ -84,7 +101,7 @@
                 <div class="card">
                     <div class="card-header">
                         <slot name="header">
-                            Add a New Work Item
+                            Edit Service Log Item
                         </slot>
                     </div>
 
@@ -93,39 +110,39 @@
                             <form action="submitForm" method="POST">
                                 <div class="mb-3">
                                     <label for="date">Date</label>
-                                    <input type="date" name="date" id="date" class="form-control" v-model="date" ref="date">
-                                    <div class="text-danger" v-if="v$.date.$error">Date field is required.</div>                              
+                                    <input type="date" name="date" id="date" class="form-control" v-model="results.service_date" ref="date">
+                                    <div class="text-danger" v-if="v$.results.service_date.$error">Date field is required.</div>                              
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="mileage">Mileage</label>
-                                    <input type="number" name="mileage" id="mileage" class="form-control" v-model="mileage" ref="mileage">
-                                    <div class="text-danger" v-if="v$.mileage.$error">Mileage field is required.</div>
+                                    <input type="number" name="mileage" id="mileage" class="form-control" v-model="results.mileage" ref="mileage">
+                                    <div class="text-danger" v-if="v$.results.mileage.$error">Mileage field is required.</div>
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="technician">Performed by</label>
-                                    <input type="text" name="technician" id="technician" class="form-control" v-model="technician" ref="technician">
-                                    <div class="text-danger" v-if="v$.technician.$error">Performed by field is required.</div>    
+                                    <input type="text" name="technician" id="technician" class="form-control" v-model="results.technician" ref="technician">
+                                    <div class="text-danger" v-if="v$.results.technician.$error">Performed by field is required.</div>    
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="cost">Cost</label>
-                                    <input type="number" step=0.01 name="cost" id="cost" class="form-control" v-model="cost" ref="cost">
-                                    <div class="text-danger" v-if="v$.cost.$error">Cost field is required.</div>    
+                                    <input type="number" step=0.01 name="cost" id="cost" class="form-control" v-model="results.cost" ref="cost">
+                                    <div class="text-danger" v-if="v$.results.cost.$error">Cost field is required.</div>    
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="service_summary">Service summary</label>
-                                    <input type="text" name="service_summary" id="service_summary" class="form-control" v-model="service_summary" ref="service_summary">
-                                    <div class="text-danger" v-if="v$.service_summary.$error">Service summary field is required.</div>                          
+                                    <input type="text" name="service_summary" id="service_summary" class="form-control" v-model="results.service_summary" ref="service_summary">
+                                    <div class="text-danger" v-if="v$.results.service_summary.$error">Service summary field is required.</div>                          
                                 </div>
 
                                 <div class="mb-3">
                                     <label for="service_details">Service details</label>
                                     <editor 
                                         api-key="no-api-key"
-                                        v-model="service_details" 
+                                        v-model="results.service_details" 
                                         ref="service_details"
                                         :init="{
                                             height: 150,
@@ -143,7 +160,7 @@
                                             Cancel
                                         </button>
 
-                                        <button class="btn btn-primary" @click.prevent="submitForm"><i class="fa fa-plus"></i> Add Work Item</button>
+                                        <button class="btn btn-primary" @click.prevent="submitForm">Update</button>
                                     </slot>
                                 </div>
                             </form>
